chore: 48 fetch product service data from firestore and allow update
As the true data source, we use firestore for all data, so it should be
fetched from there.

Aditionally, added a basic update capability in the product service that
should be able to update on firestore, as well as on Zettle. The latter
only allows change for `name` right now and might fail for variants
instead of products.