feat: introduce custom logger
Allows better tracing of logs between services, integrates properly to
GCP logging structure and extends logged information to help identify
the original cause more effectively.

Not using in all locations yet, but only present a test of usage in
proudct/server/server.go
This should be extended to all logging locations in the future, once we
agree to usage like this.