fix: change the network from host to internal
require to change the bind port in main.go file from localhost to
0.0.0.0
port mapping in docker happens to the internal ip address not the
localhost within the container. Therefore what causes the localhost to
not be accessible in the internal network but accessible when binds the
network to host.